requirements.txt
pillow
numpy
huggingface_hub
gradio_client